requirements.txt
flask
dnspython
python-whois
